Router - Now with Additional Options!
+ Added an optional array parameter to the static Router register()
method after the callable action to hold any additional options for the
route.
+ Added the ability to specify a hostname you wish to restrict the route
to. This is good if you wish to split your site/application across
multiple domains or subdomains.

* Changed the way routes are registered.
* Changed the way routes are routed.

No modification is needed to your existing registered routes, they will
continue to work as expected as if this update never happened. :)