Database Port Can Now Be a String or an Integer
After giving it some thought for a few hours, I decided that not
everyone will want to define the database port as an integer, especially
when the other connection details are strings, so I changed it up.

Now you can choose between a string or an integer for the port.