Persona Clix Engine - Now with Helpers
After thinking about it for some time, I figured it would be nice to have classes that do some little tasks for you, so I created some Helpers.

+ Added Helper: Form
+ Added Helper: Request
+ Added Beautiful Helper API Functions

Please note that I have only just lightly tested the code in this update, so use at your own risk. Due to the way that the AutoLoader works, your application won't automatically use the new code unless you tell it to, so if you prefer to have the latest update installed, but not use the new code, it is safe to do so.

Happy coding :)